+++ packages/AI/Vectors/Dupe/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log - @memberjunction/ai-vector-dupe
 
+## 5.33.0
+
+### Patch Changes
+
+- Updated dependencies [95eb27e]
+- Updated dependencies [74b0be0]
+- Updated dependencies [5cc5326]
+- Updated dependencies [7e4957d]
+  - @memberjunction/core@5.33.0
+  - @memberjunction/global@5.33.0
+  - @memberjunction/aiengine@5.33.0
+  - @memberjunction/ai-vectors@5.33.0
+  - @memberjunction/ai-vectordb@5.33.0
+  - @memberjunction/ai-vector-sync@5.33.0
+  - @memberjunction/core-entities@5.33.0
+  - @memberjunction/templates@5.33.0
+  - @memberjunction/ai@5.33.0
+
 ## 5.32.0
 
 ### Patch Changes