+++ packages/Angular/Generic/feedback/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @memberjunction/ng-feedback
 
+## 5.33.0
+
+### Patch Changes
+
+- Updated dependencies [97ed790]
+- Updated dependencies [95eb27e]
+- Updated dependencies [74b0be0]
+- Updated dependencies [5cc5326]
+- Updated dependencies [7e4957d]
+  - @memberjunction/graphql-dataprovider@5.33.0
+  - @memberjunction/core@5.33.0
+  - @memberjunction/ng-base-types@5.33.0
+  - @memberjunction/ng-shared-generic@5.33.0
+  - @memberjunction/ng-ui-components@5.33.0
+
 ## 5.32.0
 
 ### Patch Changes