+++ packages/ComponentRegistryClientSDK/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @memberjunction/component-registry-client-sdk
 
+## 5.33.0
+
+### Patch Changes
+
+- Updated dependencies [95eb27e]
+- Updated dependencies [74b0be0]
+- Updated dependencies [5cc5326]
+- Updated dependencies [7e4957d]
+- Updated dependencies [3e84676]
+  - @memberjunction/core@5.33.0
+  - @memberjunction/global@5.33.0
+  - @memberjunction/interactive-component-types@5.33.0
+
 ## 5.32.0
 
 ### Patch Changes