+++ packages/SQLGlotTS/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @memberjunction/sqlglot-ts
 
+## 5.33.0
+
 ## 5.32.0
 
 ## 5.31.0